Improve Input Nodes - Refactoring StructureItemSelectDialog (Splitting module)
